+++ NEWS.md
@@ -4,6 +4,7 @@
 
 - Added `winsorize()` function.
 - Added `winsorize` argument for Winsorized correlations.
+- Added `method = "somers"` to `correlation()`, to compute Somers's Dxy rank correlation for binary outcomes.
 - New function `to_table()`, to format output into nice markdown-tables.
 
 ## Bug fixes